Added manual servo and motor control.

+++ MultiWii.cpp
@@ -274,7 +274,7 @@ volatile int16_t failsafeCnt = 0;
 int16_t rcData[RC_CHANS];    // interval [1000;2000]
 int16_t rcSerial[8];         // interval [1000;2000] - is rcData coming from MSP
 int16_t rcCommand[4];        // interval [1000;2000] for THROTTLE and [-500;+500] for ROLL/PITCH/YAW
-uint8_t rcSerialCount = 0;   // a counter to select legacy RX when there is no more MSP rc serial data
+uint16_t rcSerialCount = 0;   // a counter to select legacy RX when there is no more MSP rc serial data
 int16_t lookupPitchRollRC[5];// lookup table for expo & RC rate PITCH+ROLL
 int16_t lookupThrottleRC[11];// lookup table for expo & mid THROTTLE
