+++ MultiWii.h
@@ -89,7 +89,7 @@ extern volatile int16_t failsafeCnt;
 extern int16_t rcData[RC_CHANS];
 extern int16_t rcSerial[8];
 extern int16_t rcCommand[4];
-extern uint8_t rcSerialCount;
+extern uint16_t rcSerialCount;
 extern int16_t lookupPitchRollRC[5];
 extern int16_t lookupThrottleRC[11];
